feat(threads): redesign activity log as a timeline rail
Winner of the prototype rounds: a continuous vertical rail where notes
are gold-anchored cards and automatic changes are quiet ticks, day
groups separated by column gap, labels beside the rail above their
entries. Prototype scaffolding removed; the full variant set is
archived on the prototype/activity-log-variants branch.
